[eclipse/xtext#1431] Limit MaxRAM when running in container
When running in a Docker container, the file
/sys/fs/cgroup/memory/memory.limit_in_bytes contains the RAM in bytes
available to the container. By default a Java 8 VM would try to limit
MaxRAM to 25% of the total available RAM on a machine, which would be
32GB on JIRO, while the container is limited to 2GB.
In a container without memory limitation the value is
9223372036854771712, which is the highest positive signed 64-bit integer
(2^63-1), rounded down to multiples of 4096.
This change sets -XX:MaxRAM in GRADLE_OPTS and MAVEN_OPTS when running
in a memory limited container.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>